Creational_Patterns
Creational design patterns deal with object creation mechanisms, trying to create objects in a manner suitable to the situation. They aim to encapsulate the knowledge about which concrete classes the system will use. This decouples the code that creates objects from the code that uses them, making it easier to maintain and extend