Structural design Pattern
Structural design patterns are concerned with object composition or, in other words, how classes and objects can be composed to form larger structures. They help ensure that if one part of a system changes, the entire system doesn't need to do so as well. These patterns focus on easing the design by identifying simple ways to realize relationships between entities.